fix(file)!: Add formatting to commit message
At work, some of us, have been trying to enforce a nice git commit
message etiquette. Specifically, the one proposed by Chris Beams. We
normally do so during code reviews, where the message of the commit is
almost as apt to be reviewed as the change it introduces.

To make our lives easier and save some time, we introduced this Python3
script to the CI pipeline which does some simple Natural Langu

BREAKING CHANGE: not actually, just tesing